Move GsonUtilities
