+++ java/src/com/ibm/streamsx/topology/generator/spl/AutonomousRegions.java
@@ -6,6 +6,7 @@ import java.util.Set;
 
 import com.google.gson.JsonObject;
 import com.ibm.streamsx.topology.builder.BVirtualMarker;
+import com.ibm.streamsx.topology.internal.gson.GsonUtilities;
 
 /**
  * Processing related to autonomous regions.
