+++ java/src/com/ibm/streamsx/topology/generator/spl/ThreadingModel.java
@@ -5,12 +5,12 @@
 package com.ibm.streamsx.topology.generator.spl;
 
 import static com.ibm.streamsx.topology.generator.spl.GraphUtilities.getUpstream;
-import static com.ibm.streamsx.topology.generator.spl.GsonUtilities.array;
-import static com.ibm.streamsx.topology.generator.spl.GsonUtilities.jboolean;
-import static com.ibm.streamsx.topology.generator.spl.GsonUtilities.jobject;
-import static com.ibm.streamsx.topology.generator.spl.GsonUtilities.jstring;
-import static com.ibm.streamsx.topology.generator.spl.GsonUtilities.nestedObject;
-import static com.ibm.streamsx.topology.generator.spl.GsonUtilities.nestedObjectCreate;
+import static com.ibm.streamsx.topology.internal.gson.GsonUtilities.array;
+import static com.ibm.streamsx.topology.internal.gson.GsonUtilities.jboolean;
+import static com.ibm.streamsx.topology.internal.gson.GsonUtilities.jobject;
+import static com.ibm.streamsx.topology.internal.gson.GsonUtilities.jstring;
+import static com.ibm.streamsx.topology.internal.gson.GsonUtilities.nestedObject;
+import static com.ibm.streamsx.topology.internal.gson.GsonUtilities.nestedObjectCreate;
 
 import java.util.Set;
