+++ java/src/com/ibm/streamsx/topology/internal/gson/GsonUtilities.java
@@ -2,7 +2,7 @@
 # Licensed Materials - Property of IBM
 # Copyright IBM Corp. 2016  
  */
-package com.ibm.streamsx.topology.generator.spl;
+package com.ibm.streamsx.topology.internal.gson;
 
 import java.util.Collection;
 import java.util.Iterator;
