+++ test/java/src/com/ibm/streamsx/topology/test/api/LowLatencyTest.java
@@ -4,8 +4,8 @@
  */
 package com.ibm.streamsx.topology.test.api;
 
-import static com.ibm.streamsx.topology.generator.spl.GsonUtilities.jstring;
-import static com.ibm.streamsx.topology.generator.spl.GsonUtilities.object;
+import static com.ibm.streamsx.topology.internal.gson.GsonUtilities.jstring;
+import static com.ibm.streamsx.topology.internal.gson.GsonUtilities.object;
 import static com.ibm.streamsx.topology.test.api.IsolateTest.getContainerId;
 import static com.ibm.streamsx.topology.test.api.IsolateTest.getContainerIdAppend;
 import static com.ibm.streamsx.topology.test.api.IsolateTest.getContainerIds;
@@ -33,8 +33,8 @@ import com.ibm.streamsx.topology.context.StreamsContextFactory;
 import com.ibm.streamsx.topology.function.ToIntFunction;
 import com.ibm.streamsx.topology.function.UnaryOperator;
 import com.ibm.streamsx.topology.generator.spl.GraphUtilities;
-import com.ibm.streamsx.topology.generator.spl.GsonUtilities;
 import com.ibm.streamsx.topology.generator.spl.SPLGenerator;
+import com.ibm.streamsx.topology.internal.gson.GsonUtilities;
 import com.ibm.streamsx.topology.test.AllowAll;
 import com.ibm.streamsx.topology.test.TestTopology;
 import com.ibm.streamsx.topology.tester.Condition;
